backend/requirements.txt
# Rwanda Report System - Python Requirements

# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.4

# Database & ORM
psycopg2-binary==2.9.9  # PostgreSQL (optional)

# File handling & Images
Pillow==10.1.0
python-magic==0.4.27

# API & HTTP
httpx==0.25.2
requests==2.31.0

# IPFS
ipfshttpclient==0.8.0a2

# Cardano Integration
blockfrost-python==0.7.0
pycardano==0.16.0

# Async Support
asgiref==3.7.1
celery==5.3.4  # Optional: for async background tasks
redis==5.0.1   # Optional: for celery broker

# Security
cryptography==41.0.7
PyJWT==2.8.1

# Utilities
python-decouple==3.8
python-dateutil==2.8.2

# Development (optional)
pytest==7.4.3
pytest-django==4.7.0
django-debug-toolbar==4.2.0
flake8==6.1.0
black==23.12.0
